% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crea_random_dfs.R
\name{crea_random_dfs}
\alias{crea_random_dfs}
\title{Crea dataframes con columnas de números aleatorios}
\usage{
crea_random_dfs(n_dfs, n_col, n_rows, interval)
}
\arguments{
\item{n_dfs}{Número de dataframes finales que queremos tener}

\item{n_col}{Número de columnas que ha de tener el df}

\item{n_rows}{Número de filas que ha de tener el df}

\item{interval}{Intervalo máximo y mínimo de los números aleatorios que rellenan el df}
}
\value{
Uno o más dfs
}
\description{
Crea dataframes con columnas de números aleatorios
}
