% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hola_mundo.R
\name{hola_mundo}
\alias{hola_mundo}
\title{Printaer un saludo al munndo}
\usage{
hola_mundo(x)
}
\arguments{
\item{x}{Una string comunmente un nombre propio}
}
\value{
un saludo acompañado del nombre introducido en \code{x}
}
\description{
Printaer un saludo al munndo
}
\examples{
hola_mundo("Paola")
hola_mundo("Álvaro")
}
