% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{msort}
\alias{msort}
\title{Sort integer vector}
\usage{
msort(vec, method = "radix")
}
\arguments{
\item{vec}{integer vector to be sorted}

\item{method}{character string specifying the algorithm used. Valid methods are
'quick' for quicksort and 'radix' for radix sort.}
}
\value{
sorted vector
}
\description{
Fast in-place sorting of an integer vector. Uses multithreading to increase sorting
speed.
}
