% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{msort_order}
\alias{msort_order}
\title{Fast-sort a vector and it's ordering}
\usage{
msort_order(vec, order = NULL)
}
\arguments{
\item{vec}{vector to be sorted}

\item{order}{a vector with integer values 1 to length(vec) that represents an ordering of vector _vec_.}
}
\value{
an integer vector containing the same values as _order_ but reordered in the the same way as _vec_.
}
\description{
Fast-sort a vector and it's ordering
}
